Split SettingsPage into different files to keep things organized.
